+++ queue_system_test.go
@@ -1,4 +1,4 @@
-// +build unit
+// +build system
 
 /* Copyright 2017 Victor Penso, Matteo Dessalvi, Rovanion Luckey
 
@@ -18,17 +18,9 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>. */
 package main
 
 import (
-	"io/ioutil"
-	"os"
 	"testing"
 )
 
-func TestParseQueueMetrics(t *testing.T) {
-	// Read the input data from a file
-	file, err := os.Open("test_data/squeue.txt")
-	if err != nil {
-		t.Fatalf("Can not open test data: %v", err)
-	}
-	data, err := ioutil.ReadAll(file)
-	t.Logf("%+v", ParseQueueMetrics(data))
+func TestQueueGetMetrics(t *testing.T) {
+	t.Logf("%+v", QueueGetMetrics())
 }
